Bootstrap sidebar and use stateParams for links
Watch for $stateChangeSuccess and use it to update which tab is
highlighted, if in fact the state has changed.

Use the tab.stateParams in the html instead of re-creating it there
with individual variable(s), in line with DRY principles.

Update applicable tab.stateParams to have the current tenantId when
currentTenantId is updated. This ensures that the tenantId dependent
links still update dynamically when currentTenantId is updated.

Use ng-href instead of ui-sref, because $scope variables inside {{ }}
blocks are not dynamically updating with ui-sref. See links below.

http://stackoverflow.com/questions/24349731/dynamically-set-the-value-of-ui-sref-angularjs
https://github.com/angular-ui/ui-router/issues/395