Nest sidebar and sidebarTarget in content view
With the sidebar alongside content in index.html, content is not
centered. This causes content without a sidebar, such as the login
page, to be off center.

Put sidebar and sideBarTarget ui-views into their own nested file, which
is loaded into the root level content ui-view. Center the content view
by making it full width, while removing the sideBar view alongside it in
index.html. Leave it up to the nested/child view to divide this space as
it sees fit.

Remove col-md-12 class from the header ui-view, since that is the
default anyway.