Move $stateParams.tenantId == currentTenantId check
Move Tenants.currentTenantId() validation check against $stateParams out
out of serversController, instead using a $stateChangeSucceess callback
on the run method of the main app module. Check that the two match if
the $stateParams.tenantId value exists, and assign that value using
Tenants.setCurrentTenantId() if it is different.

Helps keep code DRY as there are other controllers to be written with
tenantId in their route.