Use $localStorage for currentTenantId
Add call to Tenants.remove() in User.signOut() to clear local storage.

Set a watch on $stateParams.tenantId, if loading a state for a different
tenantId param than before, update with setCurrentTenantId(). Previously
the $stateChangeSuccess event was used but it would not broadcast
events when the tenantId in $stateParams changed, and the $state did
not.

Return from setCurrentTenantId() immediately, but validate that the
returned tenant exists in a freshly retrieved tenants list. If not,
broadcast a Tenants:currentTenant:invalid event. The same call to
validateTenantId(currentTenantId()) occurs when Tenants.list() retrieves
a fresh list from the server, in case a tenant has been removed.

Create a handler in HeadersController for the
tenants:CurrentToken:invalid broadcast. It will either replace the
current tenant with the one in $stateParams.tenantId, if it exists, or
otherwise pick the first tenant in the list.

Fix bug where Tenants.list() was not checking if currentTenantId() is
blank before returning a cached list. This bug was exposed when the
local storage began persisting currentTenant. currentTenant would
get cleared on sign out, whereas before the Tenants service variable it
was stored in would persist on sign out. list() would return the
cached tenant list results if someone were to login within up to ~7
seconds, the cache time. The cached return, instead of setting
currentTenant with a reasonable value, like the http return code path
would, was leaving currentTenant undefined.

Rename retrieveTenants() to uncached()