Separate current tenant, add error callback chain
The Tenants service should no longer manage who the current tenant is.
Implement CurrentTenant to handle this.

Tenants.list() should no longer set the current tenant id when it finds
it unset. Instead call setId() on $stateChangeSuccess and when
$stateParams.tenantId changes.

Implement callback chains to handle when the current tenant id is
invalid. Broadcasting events did not allow the app to define the order
in which its parts should get a crack at setting a valid current tenant
id. This allows for separation of concerns, such as the trying to use
$stateParams.tenantId at the app level, while also allowing controllers
with more detailed knowledge to insert their own callbacks before the
root app level attempt. When a callback returns a resolved promise, the
remaining callbacks get skipped, as chainAndRun() just passes the
promise along by skipping then() and only using a catch() to run the
next callback.

onErrorCallbackChain.run() is called in CurrentTenant.setId() when an id
fails to validate. If the callback chain's root function returns a
rejected promise, it is caught in setId(), and
CurrentTenant.setIdErrorHandler() is called to try and recover. It
reverts to using the existing current tenant if it validates, and
otherwise retrieves a new tenant list, choosing the first one.

Listen for $stateChangeSuccess at the root scope and if there is no
current tenant set, and if a valid user token is present, call setId()
with no params to trigger the error callback chain. Necessary because
setting the current tenant when $stateParams.tenantId changes, by
itself, won't ensure a current tenant is set every time the state
changes. Particularly when $stateParams.tenantId remains undefined,
resulting in no change detected.